Extract Version from AirbyteVersion
